Charts integrated with admin dashboard

+++ Frontend/pmp/src/component/AdminDashBoard/AdminPage.jsx
@@ -1,45 +1,46 @@
 import * as React from 'react';
-import {useEffect, useState} from 'react';
-import {DataGrid} from '@mui/x-data-grid';
+import {DataGrid}  from '@mui/x-data-grid';
 import {BottomNavigation, BottomNavigationAction, Fab} from "@mui/material";
 import PersonAddIcon from '@mui/icons-material/PersonAddAltOutlined';
+import {useEffect, useState} from "react";
 import {instance} from '../../index';
-
 
 import UserActions from './UserActions';
 import EditUser from "./EditUser";
 import {blue, red} from "@mui/material/colors";
 import EditIcon from "@mui/icons-material/Edit";
+import Signup from "../Signup";
+import {Navigate} from "react-router-dom";
 import Avatar from "@mui/material/Avatar";
 
 
 
-export default function DataTable() {
+export default function DataTable(){
 
-    const [userData, setUserData] = useState([]);
+    const [userData,setUserData]=useState([]);
 
-    const [value, setValue] = useState(0);
+    const [value, setValue]=useState(0);
 
-    const [rowSelected, setRowSelected] = useState([]);
+    const [rowSelected,setRowSelected]=useState(null);
+
+    let [userSelected,setUserSelected]=useState(null);
+
+
+    let [addUser,setaddUser]=useState(false);
 
 
     const columns = [
 
-        {field: 'id', headerName: 'ID', width: 50},
-        {
-            field: 'Photo', headerName: 'Image', width: 80,
-
+        { field: 'id', headerName: 'Id', width: 50 },
+        { field: 'Photo', headerName: 'Image',width: 80,
             renderCell: (theRow) => (<Avatar>
 
 
                 </Avatar>
             ),
-
-
         },
-        {field: 'lastname', headerName: 'Last name', width: 100},
-        {field: 'firstname', headerName: 'First name', width: 100},
-
+        { field: 'lastname', headerName: 'Last name', width: 100},
+        { field: 'firstname', headerName: 'First name', width: 100 },
         {
             field: 'fullName',
             headerName: 'Full name',
@@ -48,38 +49,31 @@ export default function DataTable() {
             valueGetter: (params) =>
                 `${params.row.firstname || ''} ${params.row.lastname || ''}`,
         },
-        {field: 'email', headerName: 'Email', width: 150, editable: true},
+        { field: 'email', headerName: 'Email', width: 150 ,  editable: true },
 
-        {field: 'password', headerName: 'Password', width: 160},
+        { field: 'password', headerName: 'Password',width: 160 },
 
 
-        {
-            field: 'deleted', headerName: 'Status', width: 130, type: 'boolean', editable: 'true'
 
-        },
-
-        {
-            field: 'Action', headerName: 'Activate', width: 120, type: 'actions',
-            renderCell: (theRow) => (<UserActions{...{theRow}}/>
+        { field: 'Action', headerName: 'Activate',width: 120 , type:'actions' ,
+            renderCell:(theRow)=>(<UserActions{...{theRow}}/>
             ),
 
         },
 
-        {
-            field: 'Edit', headerName: 'Edit', width: 120, type: 'button',
-            renderCell: (theRow) => (<EditUser {...{rowSelected}}></EditUser>
+        { field: 'Edit', headerName: 'Edit',width: 120 , type:'button' ,
+            renderCell:(theRow)=>(<EditUser {...{rowSelected}}></EditUser>
             ),
 
         },
 
-        {
-            field: 'Delete', headerName: 'Delete', width: 90, type: 'button',
-            renderCell: (theRow) => (<Fab
+        { field: 'Delete', headerName: 'Delete',width: 90 , type:'button' ,
+            renderCell:(theRow)=>(<Fab
                     sx={{
                         width: 40,
                         height: 40,
                         bgcolor: red[700],
-                        '&:hover': {bgcolor: red[900]},
+                        '&:hover': { bgcolor: red[900]},
                     }}
                     color="secondary" aria-label="delete">
                     <EditIcon type={'loading'} onClick={handleDeletion}/>
@@ -92,58 +86,57 @@ export default function DataTable() {
 
 
 
-    const handleDeletion = async () => {
+    const getUserData=async ()=>{
 
-        console.log('in delete')
-        rowSelected.forEach((row) => {
-
-            const conformation = window.confirm("Delete User " + row);
-            if (conformation) {
-                const deleteMe = async() =>{
-                    await instance.delete("/users/" + row);
-
-                    window.location.reload(false);
-                }
-                deleteMe();
-
-            }
-
-        })
-
-    }
-
-
-    const getUserData = async () => {
-
-        const data = await instance.get('/users')
+        const data=await instance.get('/users')
         setUserData(data.data);
 
     }
 
-// array of state and count on each
 
-    function count() {
+
+    function handleDeletion() {
+
+
+        if(rowSelected!=null){
+
+            rowSelected.forEach((row=>{
+
+             const conformation= window.confirm("Delete User "+row);
+
+             if (conformation){
+
+                 instance.delete('/users/'+row)
+                 window.location.reload(false);
+             }
+
+
+
+            }))
+
+        }
+
 
 
     }
-
-
-
 
 
     useEffect(() => {
 
         getUserData()
             .then(r => console.log(' axios request getUserData() successful'))
-            .catch(e => console.log("Exception thrown by axios request getUserData() "));
+            .catch(e=>console.log("Exception thrown by axios request getUserData() "));
 
 
-    }, []);
+
+    },[]);
+
 
 
     return (
 
-        <div style={{height: 600, width: '100%'}}>
+        <div style={{ height: 600, width: '100%' }}>
+
 
 
             <BottomNavigation
@@ -154,30 +147,38 @@ export default function DataTable() {
                 }}
             >
 
-                <BottomNavigationAction label="Add User" icon={<PersonAddIcon/>}/>
+                <BottomNavigationAction onClick={()=>{
 
+                    console.log('in here')
+                    setaddUser(true)}} label="Add User" icon={<PersonAddIcon />} />
+
+                {addUser && (
+                    <Navigate to="/signup" replace={true} />
+                )}
             </BottomNavigation>
 
 
+
             <DataGrid
+
 
 
                 rows={userData}
                 columns={columns}
                 pageSize={10}
                 getRowId={row => row.id}
-                rowsPerPageOptions={[5, 10, 15]}
+                rowsPerPageOptions={[5,10,15]}
                 autoPageSize={true}
 
                 checkboxSelection
                 rowSelection='single'
-                experimentalFeatures={{newEditingApi: true}}
+                experimentalFeatures={{ newEditingApi: true }}
 
 
                 sx={{
 
                     bgcolor: blue[15],
-                    marginLeft: 20
+                    marginLeft:20
                 }}
 
                 onSelectionModelChange={(ids) => {
@@ -188,10 +189,13 @@ export default function DataTable() {
                     console.log(selectedIDs);
                     setRowSelected(selectedIDs);
 
+
+
+
                 }}
-                getRowSpacing={params => ({
-                    top: params.isFirstVisible ? 0 : 5,
-                    bottom: params.isFirstVisible ? 0 : 5,
+                getRowSpacing={params =>( {
+                    top:params.isFirstVisible ? 0:5,
+                    bottom:params.isFirstVisible? 0:5,
 
                 })}
